pyproject.toml
[project]
name = "grundrisse"
version = "0.1.0"
description = "Content source management and indexing system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "edgedb>=2.0.0",
    "antidote>=2.0.0",
    "python-dotenv>=1.0.0",
    "gel>=3.1.0",
]
